Feat: 어드민 미디어 상세 페이지 컴포넌트 분리
