+++ src/components/Admin/DogGallery/GalleryViewer/MediaDisplay/styles.ts
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+export const DisplayContainer = styled.div`
+  margin-top: 2rem;
+  height: calc(100% - 18%);
+`;
+
+export const DisplayMedia = styled.img`
+  width: 100%;
+  height: 100%;
+`;
